requirements_minimal.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
